+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 
  - remove_user_from_tenant command
+ - Added support for tenant start/stop
 
 ### Updates 
  - changed handling of tenant arg in user resource